+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -81,7 +81,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-04-30T05:28:01.013Z")
+    updated: Optional[datetime] = parse_datetime("2026-04-30T08:46:57.988Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")