+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQEDQT8G93SC0MZBRYJ2S8P2"
+        "01KQES46MSASK6STVWFDTBWEGJ"
     )
     r"""The unique identifier of the data source"""